+++ app-child/src/main/java/com/fintory/child/domain/stock/controller/common/CommonStockControllerImpl.java
@@ -2,10 +2,11 @@ package com.fintory.child.domain.stock.controller.common;
 
 import com.fintory.common.api.ApiResponse;
 import com.fintory.domain.stock.dto.korean.response.StockSearchResponse;
-import com.fintory.domain.stock.service.CommonStockService;
+import com.fintory.domain.stock.service.common.CommonStockService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -13,6 +14,7 @@ import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
+@RequestMapping("/api/child/stock")
 public class CommonStockControllerImpl implements CommonStockController {
 
     private final CommonStockService commonStockService;
